pyproject.toml
[project]
name = "agenticzero"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0.0",
    "mcp>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "pyyaml>=6.0.1",
]

[dependency-groups]
dev = [
    "pyright>=1.1.402",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.0",
]
